pyproject.toml
[tool.poetry]
name = "singer-sdk"
version = "0.3.6"
description = "A framework for building Singer taps"
authors = ["Meltano Team and Contributors"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "<3.10,>=3.6.1"
pipelinewise-singer-python = "1.2.0"
backoff = ">=1.8.0,<2.0"
pendulum = "^2.1.0"
click = "~=8.0"
PyJWT = "1.7.1"
requests = "^2.25.1"
cryptography = "^3.4.6"
importlib-metadata = {version = "*", markers = "python_version < \"3.8\""}
memoization = "^0.3.2"
jsonpath-ng = "^1.5.3"
joblib = "^1.0.1"
inflection = "^0.5.1"

# Sphinx dependencies installed as optional 'docs' extras
# https://github.com/readthedocs/readthedocs.org/issues/4912#issuecomment-664002569
sphinx = {version = "^3.5.4", optional = true}
sphinx-rtd-theme = {version = "^0.5.2", optional = true}
sphinx-copybutton = {version = "^0.3.1", optional = true}
myst-parser = {version = "^0.14.0", optional = true}
simpleeval = "^0.9.10"

[tool.poetry.extras]
docs = ["sphinx", "sphinx-rtd-theme", "sphinx-copybutton", "myst-parser"]

[tool.poetry.dev-dependencies]
# snowflake-connector-python = "2.0.4" # Removed: Too many version conflicts!
pytest = "^6.1.2"
black = {version = "^20.8b1", allow-prereleases = true}
mypy = "^0.812"
cookiecutter = "^1.7.2"
PyYAML = "^5.4.1"
pydocstyle = "^5.1.1"
flake8 = "^3.9.0"
pyarrow = "^3.0.0"
pandas = ">1,<=1.2.0"
tox = "^3.23.1"
codecov = "^2.1.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# Sample CLI declaration:
# sdk-tap-countries-sample = 'singer_sdk.samples.sample_tap_countries.countries_tap:SampleTapCountries.cli'
